Add animal boolean search script in K6 benchmarks
A new K6 benchmark script has been added that allows users to run a boolean search for 'IsAlive' field in the 'Animal' template. This client script will allow users to effectively test the performance of boolean searches in the document service. It authenticates the request with an authorization token and checks if the response status is OK after the search operation.